Moved custom exceptions to umleditor directory, changed imports
