+++ .idea/workspace.xml
@@ -40,6 +40,7 @@
 }</component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\kaitl\PycharmProjects\2024sp-420-PythonA\src\umleditor\mvc_model" />
       <recent name="C:\Users\kaitl\PycharmProjects\2024sp-420-PythonA\src\umleditor" />
       <recent name="C:\Users\kaitl\PycharmProjects\2024sp-420-PythonA\src" />
     </key>
@@ -89,7 +90,7 @@
       <workItem from="1713754815078" duration="3283000" />
       <workItem from="1713985392538" duration="1647000" />
       <workItem from="1714075201122" duration="8831000" />
-      <workItem from="1714155430014" duration="730000" />
+      <workItem from="1714155430014" duration="1193000" />
     </task>
     <servers />
   </component>
