+++ src/test/test_relation.py
@@ -1,11 +1,11 @@
 from umleditor.mvc_model.relation import Relation
 from umleditor.mvc_model.entity import Entity
-from umleditor.mvc_model.custom_exceptions import CustomExceptions
+from umleditor.custom_exceptions import CustomExceptions
 
 import pytest
 
 def test_import_custom_exceptions():
-    from umleditor.mvc_model.custom_exceptions import CustomExceptions
+    from umleditor.custom_exceptions import CustomExceptions
     assert CustomExceptions
 
 def test_create_relation_success():
