+++ src/umleditor/mvc_controller/serializer.py
@@ -20,7 +20,7 @@ import umleditor.mvc_controller.controller_output as controller_output
 from umleditor.mvc_model.diagram import Diagram
 from umleditor.mvc_model.entity import Entity, UML_Method
 from umleditor.mvc_model.relation import Relation
-from umleditor.mvc_model.custom_exceptions import CustomExceptions as CE
+from umleditor.custom_exceptions import CustomExceptions as CE
 
 
 def serialize(diagram: Diagram, path: str) -> None:
